feat(macos): enhance macOS build workflow and UI components
- Added a Swift file size check and macOS tests to the build workflow for improved quality assurance.
- Updated the GlassMethodSelector to include a checkmark for the selected input method, enhancing user feedback.
- Introduced new Settings components for a more structured settings interface, including a sidebar and toolbar.
- Implemented an OnboardingFooter for better navigation during onboarding steps.
- Removed deprecated AutomationPane and streamlined the SettingsView for better maintainability.
- Updated macOS deployment target to 26.0 for compatibility.
